requirements.txt
backoff
openai
python-dotenv
